Set up automatic deployments with GitHub Actions
This commit will create in .github folder the workflows folder. In .github/workflows folder, it will create three files: main.yml, test.yml and dev.yml. These three files will configure automatic site deployment for the main branch, as well as for the test and dev branches which will later be created. This commit will also update README.md file with with the new version number of the project.