Build and style the header
This commit will create index.html file. In index.html file, basic html setup has been configured and the header element has been builded. This commit will also create resources folder and in it two more folders: css and img. In css folder, there are two files. One of them is general.css in which are, and will be, general and reusable css rules applied. The other file is style.css which is, and will be, responsible for main styling of the website. Img folder is the folder which will hold all images of the project. For now, there are two files in img folder. These are hero.jpg and logo.png. This commit will, with the help of all files that were mentioned above build and style the header of the website.