Build and style the gallery section of the website
This commit will create .prettierrc file, and add configuration for prettier exstension in Visual Studio Code. This commit will also create gallery folder in ./resources/img folder. In ./resources/img/gallery folder, there are 15 image files which will be used for the gallery section of the website. This commit will also create js folder in .resources folder, and in it script.js file. It will update index.html file in order to import GLightbox script and in order to build the gallery section. The file .resources/css/style.css is also updated in order to style the gallery section. In ./resources/js/script.js file the GLightbox API has been called to create the gallery functionality.